fix(HACBS-1553): bugs in upload sbom script
image_id was None if taken from sbom filename.
It was calculated properly, but then the wrong
variable was used.

Check for existing content_manifest and
content_manifest_components in image was wrong.
The old code worked when REST api was used, but
since I switched to GraphQL, the properties will
always exist and be None if no objects are found.

Also, split out graphql queries into a new function.
This includes improved error handling.
Before this change, in some testing I was
getting unhelpful error messages when the request
failed on a lower level (e.g. wrong query parameter).

Signed-off-by: Martin Malina <mmalina@redhat.com>