fix(RHTAPBUGS-756): set digest in repo when creating an image in Pyxis
Based on the mediaType returned by `skopeo inspect --raw`,
the create_container_image will determine if the image
is single arch or multi arch and based on that
it will populate either manifest_schema2_digest
or manifest_list_digest in the embedded
ContainerImageRepo object when creating a new
image in Pyxis.

This also removes the docker_image_digest and
image_id fields when creating the image. This is to align
with metaxor which does not populate these either.

This also affects the mechanism to find out if the image
already exists - we can no longer rely on the digest
field in the ContainerImage object, instead we have
to check the right field in the embedded repository.

Note that currently we do not really support multi arch
images in RHTAP and other changes will likely be needed
to fully support them.

Signed-off-by: Martin Malina <mmalina@redhat.com>