feat(RHTAPREL-618): add rh-push option when creating pyxis image (#76)
If set to true, the registry and repository entries in the Pyxis
Container Image object will be converted to use Red Hat's official
registry. E.g. a mapped repository
of quay.io/redhat-pending/product---my-image will be converted to use
registry registry.access.redhat.com and repository product/my-image.

Also, the image will be marked as published.

This will be used in the rh-push-to-registry-redhat-io pipeline.

Signed-off-by: Martin Malina <mmalina@redhat.com>