feat(RHTAPREL-618): create two repos when pushing to registry.redhat.io (#80)
Based on discussion with rbean and jvulgan in Slack, we should
actually create two items in ContainerImage.repositories
for images pushed to registry.redhat.io. The first one will be
the same as for other pushes. The second one will have
the registry and repository values converted (see the Python
script documentation for more details).

The quay.io entry will be used for pulling the image
for grading of the image by Clair.

Signed-off-by: Martin Malina <mmalina@redhat.com>