fix: install opm from github
Get the opm binary in the Dockerfile using make targets in the
operator-registry github repo instead of using the tarball on the
openshift mirror. This is superior because there can be delays
publishing content to the mirror. It also allows us to fix the CVE by
updating the opm binary now.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>