feat(RHTAPREL-816): add translate-delivery-repo utility
This commit adds the translate-delivery-repo to translate
quay.io/redhat-prod/product----repo into registry.redhat.io/product/repo
and quay.io/redhat-pending/product----repo into
registry.stage.redhat.io/product/repo. This translation will need to be
done in multiple places, so it should be stored as a utility in this
image instead of repeating the logic in multiple locations.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>