fix(RHTAPREL-857): handle case when there is no content_manifest yet
In the upload_sbom script, when we initially start processing
an image, there will be no associated content_manifest yet.
Previously, the content_manifest_components edge would still
return an empty array in its data field and we relied on it.

But the correct behavior is to return an error for the edge instead
and the Pyxis team deployed this change which broke our script.
They reverted the change until we address it which is happening
with this change.

This adds a check so that when there is no content_manifest,
we don't even check the components.

More error handling will be added later.

Signed-off-by: Martin Malina <mmalina@redhat.com>