fix(RELEASE-1108): logging should go to stdout
By default, logging goes to stderr in Python.
This will change it to go to stdout.
That makes it in line with our other Python scripts and also,
it will help with debugging when this script runs in the Tekton task,
because we suppress stderr there
(and only print it if the script fails).

Signed-off-by: Scott Hebert <scoheb@gmail.com>