fix(KONFLUX-4292): upload a fixed gpg key id to pyxis
To fix KONFLUX-4175, we need to upload a gpg key id to pyxis.

The right way to do this is for the SBOM itself to contain the gpg key
id. Our task here would simply copy it to pyxis.  However, to do that
fully, we'll need to get a change into both syft and cachi2, which may
take some time.

As a stopgap, this change inspects the "publisher" field of the sbom. If
it is set to Red Hat, we'll publish the Red Hat release2 key id to
pyxis.

You might ask, what if the rpms are not actually signed by Red Hat keys?
We have a separate layer of protection against that in KONFLUX-3149, but
the data about which rpms are signed by which keys will not be exposed
by that work in the way that the release process can use here.

* https://issues.redhat.com/browse/KONFLUX-4292
* https://issues.redhat.com/browse/KONFLUX-4175

This change should be reverted as a part of completing KONFLUX-4175.

Signed-off-by: Ralph Bean <rbean@redhat.com>