fix(KFLUXBUGS-1848): associate existing images with a new repository
This adds a new behavior.

If asked to create an image associated with a repository, and the image
does not exist at all - then create it, as before.

If the image does exist, but is not associated with the desired
repository, then submit a PATCH request to update it, adding the desired
repository.

Signed-off-by: Ralph Bean <rbean@redhat.com>