fix: supply manifest list digest to pyxis
We're populating containerImage entities incorrectly.

I think this has no real effect, but if you look at one that we
generate, the manifest_list_digest and manifest_schema2_digest are
always the same, but they're not supposed to be. We currently
incorrectly always set them to the value of the arch-specific image
manifest, when the point is to provide both values, the multi-arch image
index manifest as the manifest_list_digest and the arch-specific image
manifest as the manifest_schema2_digest.

Signed-off-by: Ralph Bean <rbean@redhat.com>