fix: handle none-type values more carefully
In the event that an image does not have any gzipped layers, then no
uncompressed layer information will be handed to this tool and it will
pass on the None value to pyxis. That fails pyxis' input validation,
blocking containerImage creation.

In this change, simply drop these values if they happen to be none.

Signed-off-by: Ralph Bean <rbean@redhat.com>