fix: ignore gpg-pubkey package
This rpm is never signed. It delivers the public key used to validate
all other packages.

It is present in the sbom, but the container catalog gets confused when
it is present. As an unsigned rpm, it displays it as "excluded" from
security scans.

Rather than claim that it is signed, let's exclude it from the manifest
uploaded to pyxis.

Signed-off-by: Ralph Bean <rbean@redhat.com>