fix(RELEASE-999): check for existing Pyxis image
The function that checks if the image already exists in Pyxis
would always use `repositories.manifest_schema2_digest`
in the query. But we only set that field for non-multiarch images.
For multiarch, we set `repositories.manifest_list_digest`.

So that means that for multiarch images, I don't see how this could
ever work.

The obvious fix would be to search for the right field for
the particular image. The issue with that is that for multiarch
images, we would have the same manifest list entry for each
of the arch images, so we would only create the first one and
all the rest would be skipped. So instead, let's use
the image_id field which is set to the arch specific digest
(for single arch, it's the main digest).

Signed-off-by: Martin Malina <mmalina@redhat.com>