feat(SPSTRAT-464): add wrapper for marketplacesvm
This commit introduces a wrapper for `pubtools-marketplacesvm` which
aims to be eventually used to support VM images pushes on various
cloudmarketplaces (such as AWS and Azure).

With this wrapper it will be possible to easily integrate the tooling
into Tekton CI using the Cloud Staged structure as input.

Refers to SPSTRAT-464

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>