fix(KFLUXBUGS-1846): fail get-image-architectures if skopeo failed
If skopeo inspect fails, the variable would be empty and then
the jq query would return an empty string instead of null, so
we would assume it's an oci artifact and run oras resolve.

If we were unable to fetch the image details with skopeo inspect
we should just fail the script.

Signed-off-by: Martin Malina <mmalina@redhat.com>