feat(ISV-5321): update SPDX document name value
This commit sets a name field of the component image SBOM to image
pullspec made of repository and image digest. The pullspec is parsed
from image purl.

The name is originally set in the build phase but it refers to an
internal Quay repository. In the release time we need to set it to a
public repository name that is available to customer.

JIRA: ISV-5321

Signed-off-by: Ales Raszka <araszka@redhat.com>