feat: implement updating of image tags
Before this change there was no way to rerun a release pipeline
with the intention of changing some tags - the create_container_image
would detect that the image already exists in Pyxis and has
the required repository defined and it would stop right there.

Now this scenario is supported - if the image already exists
and it includes the required repository, we check if the tags
are the same. If not, we update them to the desired list of tags.

Signed-off-by: Martin Malina <mmalina@redhat.com>