feat(RELEASE-1327): update internal-request script to support resolvers (#354)
This commit updates the utils/internal-request script to support calling
pipelines via resolvers. It keeps functionality for using request,
although the goal is to only keep it temporarily. I did want to keep
request support though in case a bug is found in the utils image before
the catalog task hitting the bug has a chance to be migrated to using
pipeline instead of request.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>