fix(RELEASE-1455): ignore purls without qualifiers in spdx sboms (#384)
The script was failing for a user when it tried to access qualifiers
for a purl dict and there were none. I discussed this with the build
team and they agreed that we should ignore such entries. In most
cases this will be a duplicate package entry anyway. More details
in Jira.

Signed-off-by: Martin Malina <mmalina@redhat.com>