feat(RELEASE-1051): get rpm summary from cachi2 annotation
Build service (via cachi2) is now able to save rpm summaries
in sboms, so try to read those and save in Pyxis.

While this feature is enabled in build pipelines by default,
there is no guarantee that it will always be there, so this
is optional - we will not fail if we cannot find it.

Signed-off-by: Martin Malina <mmalina@redhat.com>