fix(KONFLUX-6580): use set -e in get-image-architectures (#389)
If one of the skopeo or oras calls failed, the script would
continue to run and provide incorrect/empty output.

By using `set -e`, it will fail immediately if one of
the calls fails.

We could implement a retry here, but these failures
are pretty rare and we also have retry enabled on task
level, so that should be good enough for now.

Signed-off-by: Martin Malina <mmalina@redhat.com>