refactor(RELEASE-1378): remove request support in internal-request (#417)
All catalog tasks have been refactored to use git resolvers now. So, we
can drop support for the `-r <request>` field in the internal-request
script.

This makes it so the script doesn't support cluster references anymore.
If we want to do that in the future, we can using a new type of
resolver, but still not just the taskRef name, as the current -r uses.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>