fix: successful http codes in curl-with-retry (#441)
We got a report from a user showing that Jira
will return code 201 when closing an issue
and our script would consider it a failure.

Anything >= 200 and < 300 is a success.

Another potential issue could be 3xx codes,
but we don't expect those now and normally
you'd want to use -L to follow those
anyway. So we shouldn't have to worry about
that.

Signed-off-by: Martin Malina <mmalina@redhat.com>