fix: increase internal-request timeout to match
The TIMEOUT value here is the time that the parent release pipeline will
wait for the internal request to finish, and the other timeouts are the
timeouts applied to the child internalrequest pipeline.

If these are unequal, things don't make sense. We had a case today where
an internalrequest pipeline took 20m. It succeeded in the end, but only
after the parent release pipeline had given up on it. This can cause a
thundering herd problem in the downstream cluster where internal
requests pipelines execute, when parent pipelines retry and stack up
more downstream internalrequests that end up going unused.

Signed-off-by: Ralph Bean <rbean@redhat.com>