fix(RELEASE-1767): keep newlines in advisory fields
Until now, we used `>-` for multiline strings for some of the fields
in the advisory jinja template. That means that we were removing most
newlines from the input string. We shouldn't do that. We should
pass the string that the user provided on input unmodified.

With `|-`, all newlines will be preserved.

Signed-off-by: Martin Malina <mmalina@redhat.com>