feat(RELEASE-376): print pipelineRun in wait-for-internal-request (#528)
This commit modifies the wait-for-internal-request script to print out
the pipelineRun in its loop. It will only be printed out if the IR is
running or done. It is added to the existing "running" and "succeeded"
prints.

It also updates the script to write the IR and pipelineRun name to a
a file so stdout doesn't have to be read.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>