feat(RELEASE-1539): use uv to manage python dependencies
Define all Python dependencies in pyproject.toml and remove duplication
in Dockerfile. Use uv package manager for installation and management.
Uv was chosen because it's the fastest and most modern manager. Other
changes that were done to accomodate the new package manager:
- Don't define Python deps and their versions in the Dockerfile, use uv
- Declare used Python version, same as the version in the image
- Create a uv lockfile, to have pinned versions and reproducible builds
- Use uv in the Python CI pipeline
- Define weekly lockFileMaintenance, so that MintMaker would
  periodically update the lockfile
- Delete requirements.txt
- Update README

All the dependencies that were in Dockerfile are still pinned to
their specified versions in pyproject.toml.

Signed-off-by: Lubomir Gallovic <lgallovi@redhat.com>