fix: disable updates of pinned dependencies
With this config change, renovate should stop proposing updates for
pinned dependencies in pyproject.toml. It will continue to update the
lockfile, while respecting the constraints of pinned dependencies. The
reason for disabling the updates of pinned dependencies is that they
often introduce breaking changes and should only be bumped manually.

Signed-off-by: Lubomir Gallovic <lgallovi@redhat.com>