feat(RELEASE-1983): add check-embargoed-cves script
This commit adds a new script in a new cmd directory. The cmd directory
is meant to hold scripts that the release-service-catalog tasks call
directly, using the command directive. The first script in it is the
check-embargoed-cves script. This is almost an exact copy of the inline
bash code from the tekton task, with the only difference being it
accepting parameters and writing its results to locations specified by
env variables.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>