fix: id type in graphql mutation to sync rpm manifest
This new query was added just yesterday to fix a potential
issue when a rpm manifest is created in Pyxis, but not added
to the ContainerImage object in which case we do an empty update
request.

A user tried it (ocp-art) and for one of the images it hit this
corner case and tried to run the query and failed.

Signed-off-by: Martin Malina <mmalina@redhat.com>